+++ include/meson.build
@@ -450,6 +450,16 @@ if build_xorg
             'input.h',
             'inputstr.h',
             'list.h',
+            'micmap.h',
+            'micoord.h',
+            'migc.h',
+            'mi.h',
+            'miline.h',
+            'mioverlay.h',
+            'mipointer.h',
+            'mipointrst.h',
+            'mistruct.h',
+            'mizerarc.h',
             'misc.h',
             'miscstruct.h',
             'misync.h',
