+++ mi/meson.build
@@ -27,19 +27,6 @@ srcs_mi = [
     'mizerline.c',
 ]
 
-hdrs_mi = [
-  'micmap.h',
-  'micoord.h',
-  'migc.h',
-  'mi.h',
-  'miline.h',
-  'mioverlay.h',
-  'mipointer.h',
-  'mipointrst.h',
-  'mistruct.h',
-  'mizerarc.h',
-]
-
 libxserver_mi = static_library('xserver_mi',
     srcs_mi,
     include_directories: inc,
@@ -48,7 +35,3 @@ libxserver_mi = static_library('xserver_mi',
         m_dep,
     ],
 )
-
-if build_xorg
-    install_data(hdrs_mi, install_dir: xorgsdkdir)
-endif
