+++ mi/mi_priv.h
@@ -13,14 +13,14 @@
 #include "include/callback.h"
 #include "include/events.h"
 #include "include/gc.h"
+#include "include/mi.h"
+#include "include/micmap.h"
 #include "include/pixmap.h"
 #include "include/regionstr.h"
 #include "include/screenint.h"
 #include "include/scrnintstr.h"
 #include "include/validate.h"
 #include "include/window.h"
-#include "mi/mi.h"
-#include "mi/micmap.h"
 
 static inline void SetInstalledmiColormap(ScreenPtr s, ColormapPtr c) {
     dixSetPrivate(&(s)->devPrivates, micmapScrPrivateKey, c);
